Feat/kgms/v3.0.0/chenjiawei

See merge request kgcloud/kg-cloud-kgms!262

+++ kgsdk/src/main/java/com/plantdata/kgcloud/plantdata/converter/app/InfoBoxConverter.java
@@ -109,7 +109,7 @@ public class InfoBoxConverter extends BasicConverter {
 
     private static DataLinks dataLinkRspToDataLinks(DataLinkRsp dataLink) {
         DataLinks dataLinks = new DataLinks();
-        consumerIfNoNull(dataLink.getDataSetId(), a -> dataLinks.setDataSetId(a.intValue()));
+        consumerIfNoNull(dataLink.getDataSetId(), a -> dataLinks.setDataSetId(a));
         dataLinks.setDataSetTitle(dataLink.getDataSetTitle());
         dataLinks.setLinks(toListNoNull(dataLink.getLinks(), InfoBoxConverter::linkRspToLinks));
         return dataLinks;
